Change download-requirements for centos7

+++ core/src/epicli/data/common/ansible/playbooks/roles/repository/files/server/RedHat/disable-repository-server.sh
@@ -1,4 +1,9 @@
 #!/bin/bash -eu
 
-systemctl stop httpd
-systemctl disable httpd
+if [[ -z $(which systemctl) ]]; then
+  service httpd stop
+  chkconfig httpd off
+else
+  systemctl stop httpd
+  systemctl disable httpd
+fi
